+++ package.json
@@ -18,6 +18,7 @@
     "express": "^4.19.2",
     "express-session": "^1.18.0",
     "joi": "^17.13.1",
+    "jquery": "^3.7.1",
     "node": "^22.2.0",
     "nodemailer": "^6.9.13",
     "openai": "^4.47.1",
